% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KTCB.R
\name{KTCB}
\alias{KTCB}
\title{Create the special matrices K, T, B, and C}
\usage{
KTCB(n, use_sparse)
}
\arguments{
\item{n}{dimensionality of the square matrices}

\item{use_sparse}{should sparse matrices be created}
}
\value{
a list with the following fields \itemize{
\item K - the K matrix
\item T - the T matrix
\item C - the C matrix
\item B - the B matrix
}
}
\description{
create the four square matrices K, T, B, C of dimensionality n
}
\author{
christian bitter
}
